+++ commons-geometry-euclidean/src/main/java/org/apache/commons/geometry/euclidean/twod/BoundarySourceLinecaster2D.java
@@ -14,46 +14,46 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.commons.geometry.euclidean.threed;
+package org.apache.commons.geometry.euclidean.twod;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-/** Class that wraps a {@link BoundarySource3D} instance with the {@link Linecastable3D}
- * interface. This class performs a brute-force computation of the intersections of the
+/** Class that performs linecast operations against arbitrary {@link BoundarySource2D}
+ * instances. This class performs a brute-force computation of the intersections of the
  * line or line segment against all boundaries. Some data structures may support more
  * efficient algorithms and should therefore prefer those instead.
  */
-final class BoundarySourceLinecastWrapper3D implements Linecastable3D {
+final class BoundarySourceLinecaster2D implements Linecastable2D {
 
     /** The boundary source instance providing boundaries for the linecast operation. */
-    private final BoundarySource3D boundarySrc;
+    private final BoundarySource2D boundarySrc;
 
     /** Construct a new instance for linecasting against the given boundary source.
      * @param boundarySrc boundary source to linecast against.
      */
-    BoundarySourceLinecastWrapper3D(final BoundarySource3D boundarySrc) {
+    BoundarySourceLinecaster2D(final BoundarySource2D boundarySrc) {
         this.boundarySrc = boundarySrc;
     }
 
     /** {@inheritDoc} */
     @Override
-    public List<LinecastPoint3D> linecast(final Segment3D segment) {
-        final List<LinecastPoint3D> results =  getIntersectionStream(segment)
+    public List<LinecastPoint2D> linecast(final Segment segment) {
+        final List<LinecastPoint2D> results = getIntersectionStream(segment)
                 .collect(Collectors.toCollection(ArrayList::new));
 
-        LinecastPoint3D.sortAndFilter(results);
+        LinecastPoint2D.sortAndFilter(results);
 
         return results;
     }
 
     /** {@inheritDoc} */
     @Override
-    public LinecastPoint3D linecastFirst(final Segment3D segment) {
+    public LinecastPoint2D linecastFirst(final Segment segment) {
         return getIntersectionStream(segment)
-                .min(LinecastPoint3D.ABSCISSA_ORDER)
+                .min(LinecastPoint2D.ABSCISSA_ORDER)
                 .orElse(null);
     }
 
@@ -62,26 +62,26 @@ final class BoundarySourceLinecastWrapper3D implements Linecastable3D {
      * @param segment segment to intersect
      * @return a stream containing linecast intersections
      */
-    private Stream<LinecastPoint3D> getIntersectionStream(final Segment3D segment) {
+    private Stream<LinecastPoint2D> getIntersectionStream(final Segment segment) {
         return boundarySrc.boundaryStream()
                 .map(boundary -> computeIntersection(boundary, segment))
                 .filter(intersection -> intersection != null);
     }
 
-    /** Compute the intersection between a boundary facet and linecast intersecting segment. Null is
+    /** Compute the intersection between a boundary segment and linecast intersecting segment. Null is
      * returned if no intersection is discovered.
-     * @param facet facet from the boundary source
+     * @param boundary boundary from the boundary source
      * @param segment linecast segment to intersect with
      * @return the linecast intersection between the two arguments or null if there is no such
      *      intersection
      */
-    private LinecastPoint3D computeIntersection(final Facet facet, final Segment3D segment) {
-        final Vector3D intersectionPt = facet.intersection(segment);
+    private LinecastPoint2D computeIntersection(final Segment boundary, final Segment segment) {
+        final Vector2D intersectionPt = boundary.intersection(segment);
 
         if (intersectionPt != null) {
-            final Vector3D normal = facet.getPlane().getNormal();
+            final Vector2D normal = boundary.getLine().getOffsetDirection();
 
-            return new LinecastPoint3D(intersectionPt, normal, segment.getLine());
+            return new LinecastPoint2D(intersectionPt, normal, segment.getLine());
         }
 
         return null; // no intersection
