+++ commons-geometry-euclidean/src/test/java/org/apache/commons/geometry/euclidean/threed/BoundarySourceLinecaster3DTest.java
@@ -23,7 +23,7 @@ import org.apache.commons.geometry.core.precision.EpsilonDoublePrecisionContext;
 import org.apache.commons.geometry.euclidean.threed.shapes.Parallelepiped;
 import org.junit.Test;
 
-public class BoundarySourceLinecastWrapper3DTest {
+public class BoundarySourceLinecaster3DTest {
 
     private static final double TEST_EPS = 1e-10;
 
@@ -36,22 +36,22 @@ public class BoundarySourceLinecastWrapper3DTest {
     @Test
     public void testLinecast_line_simple() {
         // arrange
-        BoundarySourceLinecastWrapper3D wrapper = new BoundarySourceLinecastWrapper3D(UNIT_CUBE);
+        BoundarySourceLinecaster3D linecaster = new BoundarySourceLinecaster3D(UNIT_CUBE);
 
         // act/assert
 
         // no intersections
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returnsNothing()
             .whenGiven(Line3D.fromPointAndDirection(Vector3D.of(0, 4, 4), Vector3D.Unit.MINUS_X, TEST_PRECISION));
 
         // through center; two directions
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(Vector3D.of(0, 0.5, 0.5), Vector3D.Unit.MINUS_X)
             .and(Vector3D.of(1, 0.5, 0.5), Vector3D.Unit.PLUS_X)
             .whenGiven(Line3D.fromPointAndDirection(Vector3D.of(0.5, 0.5, 0.5), Vector3D.Unit.PLUS_X, TEST_PRECISION));
 
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(Vector3D.of(1, 0.5, 0.5), Vector3D.Unit.PLUS_X)
             .and(Vector3D.of(0, 0.5, 0.5), Vector3D.Unit.MINUS_X)
             .whenGiven(Line3D.fromPointAndDirection(Vector3D.of(0.5, 0.5, 0.5), Vector3D.Unit.MINUS_X, TEST_PRECISION));
@@ -60,10 +60,10 @@ public class BoundarySourceLinecastWrapper3DTest {
     @Test
     public void testLinecast_line_alongFace() {
         // arrange
-        BoundarySourceLinecastWrapper3D wrapper = new BoundarySourceLinecastWrapper3D(UNIT_CUBE);
+        BoundarySourceLinecaster3D linecaster = new BoundarySourceLinecaster3D(UNIT_CUBE);
 
         // act/assert
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(Vector3D.ZERO, Vector3D.Unit.MINUS_Y)
             .and(Vector3D.ZERO, Vector3D.Unit.MINUS_Z)
             .and(Vector3D.of(0, 1, 1), Vector3D.Unit.PLUS_Z)
@@ -74,19 +74,19 @@ public class BoundarySourceLinecastWrapper3DTest {
     @Test
     public void testLinecast_line_corners() {
         // arrange
-        BoundarySourceLinecastWrapper3D wrapper = new BoundarySourceLinecastWrapper3D(UNIT_CUBE);
+        BoundarySourceLinecaster3D linecaster = new BoundarySourceLinecaster3D(UNIT_CUBE);
 
         // act/assert
 
         // through single corner vertex
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(Vector3D.of(1, 1, 1), Vector3D.Unit.PLUS_Z)
             .and(Vector3D.of(1, 1, 1), Vector3D.Unit.PLUS_Y)
             .and(Vector3D.of(1, 1, 1), Vector3D.Unit.PLUS_X)
             .whenGiven(Line3D.fromPointAndDirection(Vector3D.of(1, 1, 1), Vector3D.of(1, -1, -1), TEST_PRECISION));
 
         // through two corner vertices
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(Vector3D.ZERO, Vector3D.Unit.MINUS_X)
             .and(Vector3D.ZERO, Vector3D.Unit.MINUS_Y)
             .and(Vector3D.ZERO, Vector3D.Unit.MINUS_Z)
@@ -103,10 +103,10 @@ public class BoundarySourceLinecastWrapper3DTest {
                     Facet.fromVertexLoop(Arrays.asList(Vector3D.ZERO, Vector3D.Unit.PLUS_X, Vector3D.Unit.PLUS_Y), TEST_PRECISION),
                     Facet.fromVertexLoop(Arrays.asList(Vector3D.ZERO, Vector3D.Unit.PLUS_Y, Vector3D.Unit.MINUS_X), TEST_PRECISION)
                 );
-        BoundarySourceLinecastWrapper3D wrapper = new BoundarySourceLinecastWrapper3D(src);
+        BoundarySourceLinecaster3D linecaster = new BoundarySourceLinecaster3D(src);
 
         // act/assert
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(Vector3D.of(0, 0.5, 0), Vector3D.Unit.PLUS_Z)
             .whenGiven(Line3D.fromPointAndDirection(Vector3D.of(-1, 0.5, 1), Vector3D.of(1, 0, -1), TEST_PRECISION));
     }
@@ -114,30 +114,30 @@ public class BoundarySourceLinecastWrapper3DTest {
     @Test
     public void testLinecast_segment_simple() {
         // arrange
-        BoundarySourceLinecastWrapper3D wrapper = new BoundarySourceLinecastWrapper3D(UNIT_CUBE);
+        BoundarySourceLinecaster3D linecaster = new BoundarySourceLinecaster3D(UNIT_CUBE);
 
         // act/assert
 
         // no intersections; underlying line does not intersect
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returnsNothing()
             .whenGiven(Line3D.fromPointAndDirection(Vector3D.of(0, 4, 4), Vector3D.Unit.MINUS_X, TEST_PRECISION)
                     .segment(-10, 10));
 
         // no intersections; underlying line does intersect
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returnsNothing()
             .whenGiven(Line3D.fromPointAndDirection(Vector3D.of(0.5, 0.5, 0.5), Vector3D.Unit.PLUS_X, TEST_PRECISION)
                     .segment(2, 10));
 
         // no boundaries excluded; two directions
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(Vector3D.of(0, 0.5, 0.5), Vector3D.Unit.MINUS_X)
             .and(Vector3D.of(1, 0.5, 0.5), Vector3D.Unit.PLUS_X)
             .whenGiven(Line3D.fromPointAndDirection(Vector3D.of(0.5, 0.5, 0.5), Vector3D.Unit.PLUS_X, TEST_PRECISION)
                     .segment(-10, 10));
 
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(Vector3D.of(1, 0.5, 0.5), Vector3D.Unit.PLUS_X)
             .and(Vector3D.of(0, 0.5, 0.5), Vector3D.Unit.MINUS_X)
             .whenGiven(Line3D.fromPointAndDirection(Vector3D.of(0.5, 0.5, 0.5), Vector3D.Unit.MINUS_X, TEST_PRECISION)
@@ -147,16 +147,16 @@ public class BoundarySourceLinecastWrapper3DTest {
     @Test
     public void testLinecast_segment_boundaryExcluded() {
         // arrange
-        BoundarySourceLinecastWrapper3D wrapper = new BoundarySourceLinecastWrapper3D(UNIT_CUBE);
+        BoundarySourceLinecaster3D linecaster = new BoundarySourceLinecaster3D(UNIT_CUBE);
 
         // act/assert
         Vector3D center = Vector3D.of(0.5, 0.5, 0.5);
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(Vector3D.of(1, 0.5, 0.5), Vector3D.Unit.PLUS_X)
             .whenGiven(Line3D.fromPointAndDirection(center, Vector3D.Unit.PLUS_X, TEST_PRECISION)
                     .segmentFrom(center));
 
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(Vector3D.of(1, 0.5, 0.5), Vector3D.Unit.PLUS_X)
             .whenGiven(Line3D.fromPointAndDirection(center, Vector3D.Unit.MINUS_X, TEST_PRECISION)
                     .segmentTo(center));
@@ -165,10 +165,10 @@ public class BoundarySourceLinecastWrapper3DTest {
     @Test
     public void testLinecast_segment_startEndPointsOnBoundaries() {
         // arrange
-        BoundarySourceLinecastWrapper3D wrapper = new BoundarySourceLinecastWrapper3D(UNIT_CUBE);
+        BoundarySourceLinecaster3D linecaster = new BoundarySourceLinecaster3D(UNIT_CUBE);
 
         // act/assert
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(Vector3D.of(1, 0.5, 0.5), Vector3D.Unit.PLUS_X)
             .and(Vector3D.of(0, 0.5, 0.5), Vector3D.Unit.MINUS_X)
             .whenGiven(Segment3D.fromPoints(Vector3D.of(1, 0.5, 0.5), Vector3D.of(0, 0.5, 0.5), TEST_PRECISION));
@@ -177,23 +177,23 @@ public class BoundarySourceLinecastWrapper3DTest {
     @Test
     public void testLinecast_segment_alongFace() {
         // arrange
-        BoundarySourceLinecastWrapper3D wrapper = new BoundarySourceLinecastWrapper3D(UNIT_CUBE);
+        BoundarySourceLinecaster3D linecaster = new BoundarySourceLinecaster3D(UNIT_CUBE);
 
         // act/assert
 
         // includes two intersecting boundaries
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(Vector3D.of(0, 1, 0), Vector3D.Unit.MINUS_X)
             .and(Vector3D.of(1, 1, 0), Vector3D.Unit.PLUS_X)
             .whenGiven(Segment3D.fromPoints(Vector3D.of(-1, 1, 0), Vector3D.of(2, 1, 0), TEST_PRECISION));
 
         // one intersecting boundary
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(Vector3D.of(1, 1, 0), Vector3D.Unit.PLUS_X)
             .whenGiven(Segment3D.fromPoints(Vector3D.of(0.25, 1, 0), Vector3D.of(2, 1, 0), TEST_PRECISION));
 
         // no intersecting boundary
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returnsNothing()
             .whenGiven(Segment3D.fromPoints(Vector3D.of(0.25, 1, 0), Vector3D.of(0.75, 1, 0), TEST_PRECISION));
     }
@@ -201,28 +201,28 @@ public class BoundarySourceLinecastWrapper3DTest {
     @Test
     public void testLinecast_segment_corners() {
         // arrange
-        BoundarySourceLinecastWrapper3D wrapper = new BoundarySourceLinecastWrapper3D(UNIT_CUBE);
+        BoundarySourceLinecaster3D linecaster = new BoundarySourceLinecaster3D(UNIT_CUBE);
 
         Vector3D corner = Vector3D.of(1, 1, 1);
 
         // act/assert
 
         // through corner
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(corner, Vector3D.Unit.PLUS_Z)
             .and(corner, Vector3D.Unit.PLUS_Y)
             .and(corner, Vector3D.Unit.PLUS_X)
             .whenGiven(Segment3D.fromPoints(Vector3D.of(0.5, 0.5, 0.5), Vector3D.of(2, 2, 2), TEST_PRECISION));
 
         // starts on corner
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(corner, Vector3D.Unit.PLUS_Z)
             .and(corner, Vector3D.Unit.PLUS_Y)
             .and(corner, Vector3D.Unit.PLUS_X)
             .whenGiven(Segment3D.fromPoints(corner, Vector3D.of(2, 0, 2), TEST_PRECISION));
 
         // ends on corner
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(corner, Vector3D.Unit.PLUS_Z)
             .and(corner, Vector3D.Unit.PLUS_Y)
             .and(corner, Vector3D.Unit.PLUS_X)
@@ -236,10 +236,10 @@ public class BoundarySourceLinecastWrapper3DTest {
                     Facet.fromVertexLoop(Arrays.asList(Vector3D.ZERO, Vector3D.Unit.PLUS_X, Vector3D.Unit.PLUS_Y), TEST_PRECISION),
                     Facet.fromVertexLoop(Arrays.asList(Vector3D.ZERO, Vector3D.Unit.PLUS_Y, Vector3D.Unit.MINUS_X), TEST_PRECISION)
                 );
-        BoundarySourceLinecastWrapper3D wrapper = new BoundarySourceLinecastWrapper3D(src);
+        BoundarySourceLinecaster3D linecaster = new BoundarySourceLinecaster3D(src);
 
         // act/assert
-        LinecastChecker3D.with(wrapper)
+        LinecastChecker3D.with(linecaster)
             .returns(Vector3D.of(0, 0.5, 0), Vector3D.Unit.PLUS_Z)
             .whenGiven(Segment3D.fromPoints(Vector3D.of(-1, 0.5, 1), Vector3D.of(1, 0.5, -1), TEST_PRECISION));
     }
