+++ commons-geometry-euclidean/src/test/java/org/apache/commons/geometry/euclidean/twod/BoundarySourceLinecaster2DTest.java
@@ -21,7 +21,7 @@ import org.apache.commons.geometry.core.precision.EpsilonDoublePrecisionContext;
 import org.apache.commons.geometry.euclidean.twod.shapes.Parallelogram;
 import org.junit.Test;
 
-public class BoundarySourceLinecastWrapper2DTest {
+public class BoundarySourceLinecaster2DTest {
 
     private static final double TEST_EPS = 1e-10;
 
@@ -34,22 +34,22 @@ public class BoundarySourceLinecastWrapper2DTest {
     @Test
     public void testLinecast_line_simple() {
         // arrange
-        BoundarySourceLinecastWrapper2D wrapper = new BoundarySourceLinecastWrapper2D(UNIT_SQUARE);
+        BoundarySourceLinecaster2D linecaster = new BoundarySourceLinecaster2D(UNIT_SQUARE);
 
         // act/assert
 
         // no intersections
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returnsNothing()
             .whenGiven(Line.fromPointAndDirection(Vector2D.of(0, 4), Vector2D.Unit.MINUS_X, TEST_PRECISION));
 
         // through center; two directions
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.of(0, 0.5), Vector2D.Unit.MINUS_X)
             .and(Vector2D.of(1, 0.5), Vector2D.Unit.PLUS_X)
             .whenGiven(Line.fromPointAndDirection(Vector2D.of(0.5, 0.5), Vector2D.Unit.PLUS_X, TEST_PRECISION));
 
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.of(1, 0.5), Vector2D.Unit.PLUS_X)
             .and(Vector2D.of(0, 0.5), Vector2D.Unit.MINUS_X)
             .whenGiven(Line.fromPointAndDirection(Vector2D.of(0.5, 0.5), Vector2D.Unit.MINUS_X, TEST_PRECISION));
@@ -58,10 +58,10 @@ public class BoundarySourceLinecastWrapper2DTest {
     @Test
     public void testLinecast_line_alongFace() {
         // arrange
-        BoundarySourceLinecastWrapper2D wrapper = new BoundarySourceLinecastWrapper2D(UNIT_SQUARE);
+        BoundarySourceLinecaster2D linecaster = new BoundarySourceLinecaster2D(UNIT_SQUARE);
 
         // act/assert
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.of(0, 1), Vector2D.Unit.MINUS_X)
             .and(Vector2D.of(1, 1), Vector2D.Unit.PLUS_X)
             .whenGiven(Line.fromPointAndDirection(Vector2D.of(0, 1), Vector2D.Unit.PLUS_X, TEST_PRECISION));
@@ -70,18 +70,18 @@ public class BoundarySourceLinecastWrapper2DTest {
     @Test
     public void testLinecast_line_corners() {
         // arrange
-        BoundarySourceLinecastWrapper2D wrapper = new BoundarySourceLinecastWrapper2D(UNIT_SQUARE);
+        BoundarySourceLinecaster2D linecaster = new BoundarySourceLinecaster2D(UNIT_SQUARE);
 
         // act/assert
 
         // through single corner vertex
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.of(1, 1), Vector2D.Unit.PLUS_Y)
             .and(Vector2D.of(1, 1), Vector2D.Unit.PLUS_X)
             .whenGiven(Line.fromPointAndDirection(Vector2D.of(0, 2), Vector2D.of(1, -1), TEST_PRECISION));
 
         // through two corner vertices
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.ZERO, Vector2D.Unit.MINUS_X)
             .and(Vector2D.ZERO, Vector2D.Unit.MINUS_Y)
             .and(Vector2D.of(1, 1), Vector2D.Unit.PLUS_Y)
@@ -96,10 +96,10 @@ public class BoundarySourceLinecastWrapper2DTest {
                     Segment.fromPoints(Vector2D.of(-1, -1), Vector2D.ZERO, TEST_PRECISION),
                     Segment.fromPoints(Vector2D.ZERO, Vector2D.of(1, 1), TEST_PRECISION)
                 );
-        BoundarySourceLinecastWrapper2D wrapper = new BoundarySourceLinecastWrapper2D(src);
+        BoundarySourceLinecaster2D linecaster = new BoundarySourceLinecaster2D(src);
 
         // act/assert
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.ZERO, Vector2D.Unit.from(1, -1))
             .whenGiven(Line.fromPointAndDirection(Vector2D.ZERO, Vector2D.Unit.PLUS_X, TEST_PRECISION));
     }
@@ -107,30 +107,30 @@ public class BoundarySourceLinecastWrapper2DTest {
     @Test
     public void testLinecast_segment_simple() {
         // arrange
-        BoundarySourceLinecastWrapper2D wrapper = new BoundarySourceLinecastWrapper2D(UNIT_SQUARE);
+        BoundarySourceLinecaster2D linecaster = new BoundarySourceLinecaster2D(UNIT_SQUARE);
 
         // act/assert
 
         // no intersections; underlying line does not intersect
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returnsNothing()
             .whenGiven(Line.fromPointAndDirection(Vector2D.of(0, 4), Vector2D.Unit.MINUS_X, TEST_PRECISION)
                     .segment(-10, 10));
 
         // no intersections; underlying line does intersect
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returnsNothing()
             .whenGiven(Line.fromPointAndDirection(Vector2D.of(0.5, 0.5), Vector2D.Unit.PLUS_X, TEST_PRECISION)
                     .segment(2, 10));
 
         // no boundaries excluded; two directions
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.of(0, 0.5), Vector2D.Unit.MINUS_X)
             .and(Vector2D.of(1, 0.5), Vector2D.Unit.PLUS_X)
             .whenGiven(Line.fromPointAndDirection(Vector2D.of(0.5, 0.5), Vector2D.Unit.PLUS_X, TEST_PRECISION)
                     .segment(-10, 10));
 
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.of(1, 0.5), Vector2D.Unit.PLUS_X)
             .and(Vector2D.of(0, 0.5), Vector2D.Unit.MINUS_X)
             .whenGiven(Line.fromPointAndDirection(Vector2D.of(0.5, 0.5), Vector2D.Unit.MINUS_X, TEST_PRECISION)
@@ -140,16 +140,16 @@ public class BoundarySourceLinecastWrapper2DTest {
     @Test
     public void testLinecast_segment_boundaryExcluded() {
         // arrange
-        BoundarySourceLinecastWrapper2D wrapper = new BoundarySourceLinecastWrapper2D(UNIT_SQUARE);
+        BoundarySourceLinecaster2D linecaster = new BoundarySourceLinecaster2D(UNIT_SQUARE);
 
         // act/assert
         Vector2D center = Vector2D.of(0.5, 0.5);
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.of(1, 0.5), Vector2D.Unit.PLUS_X)
             .whenGiven(Line.fromPointAndDirection(center, Vector2D.Unit.PLUS_X, TEST_PRECISION)
                     .segmentFrom(center));
 
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.of(1, 0.5), Vector2D.Unit.PLUS_X)
             .whenGiven(Line.fromPointAndDirection(center, Vector2D.Unit.MINUS_X, TEST_PRECISION)
                     .segmentTo(center));
@@ -158,10 +158,10 @@ public class BoundarySourceLinecastWrapper2DTest {
     @Test
     public void testLinecast_segment_startEndPointsOnBoundaries() {
         // arrange
-        BoundarySourceLinecastWrapper2D wrapper = new BoundarySourceLinecastWrapper2D(UNIT_SQUARE);
+        BoundarySourceLinecaster2D linecaster = new BoundarySourceLinecaster2D(UNIT_SQUARE);
 
         // act/assert
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.of(1, 0.5), Vector2D.Unit.PLUS_X)
             .and(Vector2D.of(0, 0.5), Vector2D.Unit.MINUS_X)
             .whenGiven(Segment.fromPoints(Vector2D.of(1, 0.5), Vector2D.of(0, 0.5), TEST_PRECISION));
@@ -170,23 +170,23 @@ public class BoundarySourceLinecastWrapper2DTest {
     @Test
     public void testLinecast_segment_alongFace() {
         // arrange
-        BoundarySourceLinecastWrapper2D wrapper = new BoundarySourceLinecastWrapper2D(UNIT_SQUARE);
+        BoundarySourceLinecaster2D linecaster = new BoundarySourceLinecaster2D(UNIT_SQUARE);
 
         // act/assert
 
         // includes two intersecting boundaries
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.of(0, 1), Vector2D.Unit.MINUS_X)
             .and(Vector2D.of(1, 1), Vector2D.Unit.PLUS_X)
             .whenGiven(Segment.fromPoints(Vector2D.of(-1, 1), Vector2D.of(2, 1), TEST_PRECISION));
 
         // one intersecting boundary
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.of(1, 1), Vector2D.Unit.PLUS_X)
             .whenGiven(Segment.fromPoints(Vector2D.of(0.25, 1), Vector2D.of(2, 1), TEST_PRECISION));
 
         // no intersecting boundary
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returnsNothing()
             .whenGiven(Segment.fromPoints(Vector2D.of(0.25, 1), Vector2D.of(0.75, 1), TEST_PRECISION));
     }
@@ -194,24 +194,24 @@ public class BoundarySourceLinecastWrapper2DTest {
     @Test
     public void testLinecast_segment_corners() {
         // arrange
-        BoundarySourceLinecastWrapper2D wrapper = new BoundarySourceLinecastWrapper2D(UNIT_SQUARE);
+        BoundarySourceLinecaster2D linecaster = new BoundarySourceLinecaster2D(UNIT_SQUARE);
 
         // act/assert
 
         // through corner
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.of(1, 1), Vector2D.Unit.PLUS_Y)
             .and(Vector2D.of(1, 1), Vector2D.Unit.PLUS_X)
             .whenGiven(Segment.fromPoints(Vector2D.of(0, 2), Vector2D.of(2, 0), TEST_PRECISION));
 
         // starts on corner
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.of(1, 1), Vector2D.Unit.PLUS_Y)
             .and(Vector2D.of(1, 1), Vector2D.Unit.PLUS_X)
             .whenGiven(Segment.fromPoints(Vector2D.of(1, 1), Vector2D.of(2, 0), TEST_PRECISION));
 
         // ends on corner
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.of(1, 1), Vector2D.Unit.PLUS_Y)
             .and(Vector2D.of(1, 1), Vector2D.Unit.PLUS_X)
             .whenGiven(Segment.fromPoints(Vector2D.of(0, 2), Vector2D.of(1, 1), TEST_PRECISION));
@@ -224,10 +224,10 @@ public class BoundarySourceLinecastWrapper2DTest {
                     Segment.fromPoints(Vector2D.of(-1, -1), Vector2D.ZERO, TEST_PRECISION),
                     Segment.fromPoints(Vector2D.ZERO, Vector2D.of(1, 1), TEST_PRECISION)
                 );
-        BoundarySourceLinecastWrapper2D wrapper = new BoundarySourceLinecastWrapper2D(src);
+        BoundarySourceLinecaster2D linecaster = new BoundarySourceLinecaster2D(src);
 
         // act/assert
-        LinecastChecker2D.with(wrapper)
+        LinecastChecker2D.with(linecaster)
             .returns(Vector2D.ZERO, Vector2D.Unit.from(1, -1))
             .whenGiven(Segment.fromPoints(Vector2D.ZERO, Vector2D.Unit.PLUS_X, TEST_PRECISION));
     }
